file-converter-service/requirements.txt
Flask
flask-cors
requests
PyMuPDF
python-docx
pymongo

